# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StudentInput {
  firstName: string;
  lastName: string;
  admissionNo: string;
  dateOfBirth: string;
  className: string;
  section: string;
  rollNumber: number;
  parentIds: string[];
  active: boolean;
}

export interface Subject {
  id: string;
  name: string;
  teacher: string;
}

export interface Attendance {
  present: number;
  absent: number;
  percentage: number;
}

export interface Result {
  subject: string;
  score: number;
  grade: string;
}

export interface FeeSummary {
  total: number;
  paid: number;
  due: number;
  nextDue: string;
}

export interface ParentLink {
  id: string;
  name: string;
  relationship: string;
  phone: string;
}

export type Student = StudentInput & {
  id: string;
  subjects: Subject[];
  attendance: Attendance;
  results: Result[];
  fees: FeeSummary;
  parents: ParentLink[];
};

export type StudentProfile = Student;

export interface ParentInput {
  firstName: string;
  lastName: string;
  relationship: string;
  email: string;
  phone: string;
  occupation: string;
  studentIds: string[];
}

export interface StudentLink {
  id: string;
  name: string;
  className: string;
  section: string;
  feesDue: number;
}

export type Parent = ParentInput & {
  id: string;
  students: StudentLink[];
};

export interface AdminSummary {
  totalStudents: number;
  activeStudents: number;
  totalParents: number;
  feesCollected: number;
  feesDue: number;
  attendanceRate: number;
}

export type ListStudentsParams = {
search?: string;
status?: ListStudentsStatus;
};

export type ListStudentsStatus = typeof ListStudentsStatus[keyof typeof ListStudentsStatus];


export const ListStudentsStatus = {
  all: 'all',
  active: 'active',
  inactive: 'inactive',
} as const;

export type ListParentsParams = {
search?: string;
};