# lib/api-zod/src/generated/types/parent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParentInput } from './parentInput';
import type { StudentLink } from './studentLink';

export type Parent = ParentInput & {
  id: string;
  students: StudentLink[];
};