# lib/api-zod/src/generated/types/parentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ParentInput {
  firstName: string;
  lastName: string;
  relationship: string;
  email: string;
  phone: string;
  occupation: string;
  studentIds: string[];
}