# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Attendance } from './attendance';
import type { FeeSummary } from './feeSummary';
import type { ParentLink } from './parentLink';
import type { Result } from './result';
import type { StudentInput } from './studentInput';
import type { Subject } from './subject';

export type Student = StudentInput & {
  id: string;
  subjects: Subject[];
  attendance: Attendance;
  results: Result[];
  fees: FeeSummary;
  parents: ParentLink[];
};