# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  firstName: string;
  lastName: string;
  admissionNo: string;
  dateOfBirth: string;
  className: string;
  section: string;
  rollNumber: number;
  parentIds: string[];
  active: boolean;
}