# lib/api-zod/src/generated/types/studentLink.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentLink {
  id: string;
  name: string;
  className: string;
  section: string;
  feesDue: number;
}