# lib/api-zod/src/generated/types/studentProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Student } from './student';

export type StudentProfile = Student;