# lib/api-zod/src/generated/types/subject.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Subject {
  id: string;
  name: string;
  teacher: string;
}